lxd/logs: Don't allow removing lxc.conf or lxc.log
Those files are used by the container runtime.

+++ lxd/container_logs.go
@@ -93,6 +93,10 @@ func containerLogDelete(d *Daemon, r *http.Request) Response {
 		return BadRequest(fmt.Errorf("log file name %s not valid", file))
 	}
 
+	if file == "lxc.log" || file == "lxc.conf" {
+		return BadRequest(fmt.Errorf("lxc.log and lxc.conf may not be deleted"))
+	}
+
 	return SmartError(os.Remove(shared.LogPath(name, file)))
 }
